Move tidy to commands/tidy
