+++ src/cli/commands/mod.rs
@@ -4,11 +4,14 @@ mod install;
 mod lock;
 mod pip;
 mod show;
+mod tidy;
 mod venv;
+
 pub use develop::develop;
 pub use init::init;
 pub use install::install;
 pub use lock::{bump_in_lock, metadata, update_lock};
 pub use pip::{get_frozen_deps, install_editable, install_editable_with_constraint, upgrade_pip};
 pub use show::{show_deps, show_outdated, show_venv_bin_path, show_venv_path};
+pub use tidy::tidy;
 pub use venv::{clean_venv, create_venv, ensure_venv, expect_venv};
